+++ src/frontend/Orion/Administrator/ListsConfigManagement/utils/getSelectedDashboardTool.js
@@ -1,0 +1,4 @@
+import queryString from 'query-string'
+
+export const getSelectedDashboardTool = (location) =>
+  ((location.search && queryString.parse(location.search)) || {}).dashboardTool